+++ unstructured/ingest/runner/gcs.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from unstructured.ingest.interfaces import ProcessorConfigs, StandardConnectorConfig
 from unstructured.ingest.logger import ingest_log_streaming_init, logger
-from unstructured.ingest.process import process_documents
+from unstructured.ingest.processor import process_documents
 from unstructured.ingest.runner.utils import update_download_dir_remote_url
